Clean up C usage to compile warning-free with -std=c89/-std=c99
Build the whole tree (libXm, libMrm, mwm, uil, xmbind, wml, demos) from
clean at -O2 with -Wall -Wextra under -std=c89 and -std=c99: zero
warnings, zero errors.

Fixes:
- Replace ~150 uses of BSD-only 'caddr_t' with XtPointer/XPointer
  (caddr_t is not visible under __STRICT_ANSI__; XFindContext call
  sites need XPointer *).
- Replace 4 C++-style // comments in lib/Xm (illegal in ISO C90).
- Ship the canonical 2x2 'gray' stipple in bitmaps/ and add
   to Xm's include path; drop the xbitmaps-package
  dependency of I18List.c/IconButton.c.
- Include <string.h> unconditionally in XpmI.h (was SYSV-gated, so all
  22 Xpm files used strcpy/strlen with implicit declarations).
- Add missing '#include <config.h>' before libc headers in files using
  snprintf/isascii/fileno (Paned, FontS, IconBox, Png, demos).
- Define POSIX feature-test macros via AH_TOP/config.h.in so
  isascii/fileno/C99 snprintf are declared under strict -std modes.
- tools/wml: flex scanners include config.h; add -D_DEFAULT_SOURCE and
  -Wno-unused-function via per-object rules; wmluiltok gains a real
  main() (previously linked by accident against libfl's main);
  'make clean' no longer deletes the distributed Uil.c/Uil.h.
- lib/Xm/Xmos_r.h: honor XNO_MTSAFE_DIRENTAPI so _XReaddir uses the
  readdir()-based wrapper instead of deprecated readdir_r (define set
  in config.h).

Real bugs found by the warnings:
- DataF.c df_IsWSpace scanned white_space[num_entries] (one past end).
- XmRenderT.c ReadToken returned the address of a stack local.
- DropDown.c XmDropDownGetChild switch had no 'case' keywords: four
  bare labels, so the switch never matched.
- XmString.c strcpy from a possibly-NULL source pointer.
- UilSemVal/UilSymDef: sym_m_has_cycle used bit 2 (colliding with
  sym_m_cycle_checked) and compared == 1, always false.
- TabList: uninitialized XmTabAttributeRec passed to Modify.
- Uninitialized-variable paths in Notebook, TextF, TextIn, DataF,
  ResConvert, GrabShell, WmEvent, WmError (noreturn annotation for
  ExitWM/WmXtErrorHandler), WmResParse.
- Outline.c: missing semicolons after _XmProcessLock()/Unlock() macros
  (invalid code that only compiled by accident).
- motifshell: sprintf(NULL, ...) crash path; tempnam replaced with a
  mkstemp-based helper in demos/programs/workspace/xrmLib.c.
- Dead code removal (unused vars/functions) across lib/Xm, mwm, uil,
  wml and demos; explicit /* FALLTHRU */ markers on intended
  switch fall-throughs.

Documented build-wide suppressions in configure.ac (idiomatic Xt
patterns, commented): -Wno-missing-field-initializers (zero-filled
class records), -Wno-cast-function-type (ABI-compatible callback
signatures), -Wno-sequence-point (_XmIsFastSubclass scratch pointer),
-Wno-format-truncation/-Wno-format-overflow (snprintf clipping by
design), -Wno-redundant-decls, plus pre-existing sign/parameter
suppressions.  --enable-extra-warnings enables the -Wextra set.

Note: regenerated wmllex.c/wmluiltok.c with flex 2.6.4 (matches the
.wl prologue fix); wmlparse.c/wmlparse.h remain the shipped byacc
output.

+++ demos/programs/Ext18List/util-c.c
@@ -1412,17 +1412,17 @@ XtPointer CONVERT
 	switch(toVal.size)
 	{
 	case 1:
-	    val = (XTPOINTER)(unsigned)(*(char*)toVal.addr);
+	    val = (XTPOINTER)(unsigned long)(*(char*)toVal.addr);
 	    break;
 	case 2:
-	    val = (XTPOINTER)(unsigned)(*(short*)toVal.addr);
+	    val = (XTPOINTER)(unsigned long)(*(short*)toVal.addr);
 	    break;
 	case 4:
-	    val = (XTPOINTER)(unsigned)(*(int*)toVal.addr);
+	    val = (XTPOINTER)(unsigned long)(*(int*)toVal.addr);
 	    break;
 	case 8:
 	default:
-	    val = (XTPOINTER)(unsigned)(*(long*)toVal.addr);
+	    val = (XTPOINTER)(unsigned long)(*(long*)toVal.addr);
 	    break;
 	}
     }
